==> R/02_prepare_paths.R <==
## ---- prepare_paths
if (!dir.exists("../data")) dir.create("../data")
if (!dir.exists("../data/processed")) dir.create("../data/processed")
if (!dir.exists("../data/modelled")) dir.create("../data/modelled")
## ----end
